+++ app/Http/Controllers/TimelineController.php
@@ -80,6 +80,8 @@ class TimelineController extends Controller
    
    public function create()
    {
+      $this->authorize('Admin');
+
       //  Return the form to create a new event
       return view('timeline.create');
    }
@@ -90,6 +92,8 @@ class TimelineController extends Controller
    
    public function store(Request $request)
    {
+      $this->authorize('Admin');
+
       // Open the DB ready for new event
       $new_event = new Timeline;
       
@@ -129,6 +133,8 @@ class TimelineController extends Controller
    
    public function edit($id)
    {
+      $this->authorize('Admin');
+
       // Get required data elements
       $timeline_event = Timeline::find($id);
       
@@ -142,6 +148,7 @@ class TimelineController extends Controller
    
    public function update(Request $request, $id)
    {
+      $this->authorize('Admin');
       // Get required data elements
       $edit_event = Timeline::find($id);
       
@@ -170,6 +177,7 @@ class TimelineController extends Controller
    
    public function destroy($id)
    {
+      $this->authorize('Admin');
       // Destroy the timeline event
       Timeline::destroy($id);
